Add recycling view to appointments list
