+++ src/main/java/seedu/contax/ui/MainWindow.java
@@ -17,6 +17,7 @@ import seedu.contax.logic.Logic;
 import seedu.contax.logic.commands.CommandResult;
 import seedu.contax.logic.commands.exceptions.CommandException;
 import seedu.contax.logic.parser.exceptions.ParseException;
+import seedu.contax.ui.appointment.AppointmentListPanel;
 import seedu.contax.ui.onboarding.OnboardingPrompt;
 
 /**
