+++ src/main/java/seedu/contax/ui/appointment/AppointmentListPanel.java
@@ -1,4 +1,4 @@
-package seedu.contax.ui;
+package seedu.contax.ui.appointment;
 
 import java.util.logging.Logger;
 
@@ -9,6 +9,9 @@ import javafx.scene.control.ListView;
 import javafx.scene.layout.Region;
 import seedu.contax.commons.core.LogsCenter;
 import seedu.contax.model.chrono.ScheduleItem;
+import seedu.contax.ui.UiPart;
+import seedu.contax.ui.appointment.factory.AppointmentListRowFactory;
+import seedu.contax.ui.appointment.factory.AppointmentListRow;
 
 /**
  * Panel containing a list of appointments.
@@ -20,14 +23,14 @@ public class AppointmentListPanel extends UiPart<Region> {
     @FXML
     private ListView<ScheduleItem> appointmentListView;
 
-    private final AppointmentListCardFactory cardFactory;
+    private final AppointmentListRowFactory cardFactory;
 
     /**
      * Creates a {@code AppointmentListPanel} with the given appointment {@code ObservableList}.
      */
     public AppointmentListPanel(ObservableList<ScheduleItem> appointmentList) {
         super(FXML);
-        this.cardFactory = new AppointmentListCardFactory(appointmentList);
+        this.cardFactory = new AppointmentListRowFactory(appointmentList);
         appointmentListView.setItems(appointmentList);
         appointmentListView.setCellFactory(listView -> new AppointmentListPanel.AppointmentListViewCell());
     }
@@ -37,6 +40,13 @@ public class AppointmentListPanel extends UiPart<Region> {
      * an {@code AppointmentCard}.
      */
     class AppointmentListViewCell extends ListCell<ScheduleItem> {
+
+        private final AppointmentListRow row;
+
+        AppointmentListViewCell() {
+            row = cardFactory.createRow();
+        }
+
         @Override
         protected void updateItem(ScheduleItem scheduleItem, boolean empty) {
             super.updateItem(scheduleItem, empty);
@@ -45,7 +55,7 @@ public class AppointmentListPanel extends UiPart<Region> {
                 setGraphic(null);
                 setText(null);
             } else {
-                setGraphic(cardFactory.createCard(scheduleItem, getIndex()));
+                setGraphic(cardFactory.updateRow(row, scheduleItem, getIndex()));
             }
         }
     }
