+++ src/main/java/seedu/contax/ui/appointment/factory/AppointmentListRowFactory.java
@@ -1,28 +1,28 @@
-package seedu.contax.ui;
+package seedu.contax.ui.appointment.factory;
 
 import java.util.ArrayList;
 
 import javafx.beans.InvalidationListener;
 import javafx.collections.ObservableList;
 import javafx.scene.Node;
-import seedu.contax.model.appointment.Appointment;
 import seedu.contax.model.appointment.AppointmentSlot;
 import seedu.contax.model.chrono.ScheduleItem;
 
 /**
- * Creates elements for the {@code AppointmentListPanel}.
+ * Creates rows for the {@code AppointmentListPanel} and transparently performs the translation from row's
+ * ListView index to its displayed index.
  */
-public class AppointmentListCardFactory {
+public class AppointmentListRowFactory {
 
     private final ObservableList<ScheduleItem> scheduleItemList;
     private final ArrayList<Integer> indexMap;
 
     /**
-     * Constructs an AppointmentListCardFactory.
+     * Constructs an AppointmentListRowFactory.
      *
      * @param scheduleItemList The list of models that the factory is creating cards for.
      */
-    public AppointmentListCardFactory(ObservableList<ScheduleItem> scheduleItemList) {
+    public AppointmentListRowFactory(ObservableList<ScheduleItem> scheduleItemList) {
         this.scheduleItemList = scheduleItemList;
         this.indexMap = new ArrayList<>(scheduleItemList.size());
         populateIndexMap();
@@ -30,20 +30,22 @@ public class AppointmentListCardFactory {
     }
 
     /**
-     * Creates UI {@code Node} objects for use in the {@code AppointmentListPanel}.
+     * Creates a new {@code AppointmentListRow} that represents a row in {@code AppointmentListPanel}.
+     */
+    public AppointmentListRow createRow() {
+        return new AppointmentListRow();
+    }
+
+    /**
+     * Updates the supplied {@code row} with the data in the {@code ScheduleItem} model and the list index.
      *
-     * @param scheduleItem The ScheduleItem for which a row in the list should be created.
+     * @param row The row to update.
+     * @param scheduleItem The ScheduleItem that the supplied row should be updated to.
      * @param index The 0-based index of this row in the ListView.
      * @return A row that can be added as a child to {@code AppointmentListPanel}.
      */
-    public Node createCard(ScheduleItem scheduleItem, int index) {
-        if (scheduleItem instanceof Appointment) {
-            return new AppointmentCard((Appointment) scheduleItem, indexMap.get(index) + 1).getRoot();
-        } else if (scheduleItem instanceof AppointmentSlot) {
-            return new AppointmentSlotCard((AppointmentSlot) scheduleItem).getRoot();
-        } else {
-            return null;
-        }
+    public Node updateRow(AppointmentListRow row, ScheduleItem scheduleItem, int index) {
+        return row.updateModel(scheduleItem, indexMap.get(index) + 1);
     }
 
     /**
